move ImageWrapper_to_GLTexture.h to Image Presentation Service
