+++ frontend/services/image_presentation/ImageWrapper_to_GLTexture.hpp
